final source file naming fix
