+++ arithmetic_analysis/newton_method.py
