+++ arithmetic_analysis/newton_raphson_method.py
