+++ boolean_algebra/quine_mc_cluskey/quine_mc_cluskey.py
