+++ data_structures/binary_tree/fenwick_tree.py
