+++ data_structures/binary_tree/lazy_segment_tree.py
