+++ data_structures/binary_tree/segment_tree.py
