+++ data_structures/linked_list/singly_linked_list.py
