+++ maths/simpson_rule.py
