+++ maths/trapezoidal_rule.py
