+++ strings/knuth_morris_pratt.py
